+++ events/event.h
@@ -60,52 +60,4 @@ namespace QMatrixClient
     using Events = QVector<Event*>;
 
     Events eventsFromJson(const QJsonArray& json);
-
-    /**
-     * @brief Lookup a value by a key in a varargs list
-     *
-     * The below overloaded function template takes the value of its first
-     * argument (selector) as a key and searches for it in the key-value map
-     * passed in a varargs list (every next pair of arguments forms a key-value
-     * pair). If a match is found, the respective value is returned; otherwise,
-     * the last value (fallback) is returned.
-     *
-     * All options should be of the same type or implicitly castable to the
-     * type of the first option. Note that pointers to methods of different
-     * classes are of different object types, in particular.
-     *
-     * Below is an example of usage to select a parser depending on contents of
-     * a JSON object:
-     * {@code
-     *  auto parser = lookup(obj.value["type"].toString(),
-     *                      "type1", fn1,
-     *                      "type2", fn2,
-     *                      fallbackFn);
-     *  parser(obj);
-     * }
-     *
-     * The implementation is based on tail recursion; every recursion step
-     * removes 2 arguments (match and option). There's no selector value for the
-     * fallback option (the last one); therefore, the total number of lookup()
-     * arguments should be even: selector + n key-value pairs + fallback
-     *
-     * @note Beware of calling lookup() with a <code>const char*</code> selector
-     * (the first parameter) - most likely it won't do what you expect because
-     * of shallow comparison.
-     */
-    template <typename ValueT, typename SelectorT, typename KeyT, typename... Ts>
-    ValueT lookup(SelectorT selector, KeyT key, ValueT value, Ts... remainingMapping)
-    {
-        if( selector == key )
-            return value;
-
-        // Drop the failed key-value pair and recurse with 2 arguments less.
-        return lookup(selector, remainingMapping...);
-    }
-
-    template <typename SelectorT, typename ValueT>
-    ValueT lookup(SelectorT/*unused*/, ValueT fallback)
-    {
-        return fallback;
-    }
 }
